% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.lib.R
\name{availableTransforms}
\alias{availableTransforms}
\title{Get a list of available transforms.}
\usage{
availableTransforms()
}
\value{
A character vector of transform names.
}
\description{
Get a list of available transforms.
}
\examples{
 availableTransforms()
}
