% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{cmd_str_fmt}
\alias{cmd_str_fmt}
\title{Format a TCL Command Procedure for SOLAR}
\usage{
cmd_str_fmt(tcl_proc_name, cmd_fn_body)
}
\arguments{
\item{tcl_proc_name}{Character. The name of the TCL procedure.}

\item{cmd_fn_body}{Character. The body of the TCL command.}
}
\value{
A character vector representing the full TCL function.
}
\description{
Creates a TCL procedure string with the correct header and body.
}
