% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{CreateEvdData}
\alias{CreateEvdData}
\title{R6 Class CreateEvdData}
\description{
TODO: Add description
}
\details{
From the Solar manual:

create_evd_data performs an EVD on the loaded pedigree for gwas or. gpu_gwas
commands. This is useful for a data set with a large number of subjects.

USAGE:
create_evd_data --o \if{html}{\out{<output base filename>}}
--plink \if{html}{\out{<plink set base filename>}} --use_covs

Prior to running the command select the trait that you plan to run gwas,
gpu_gwas, or gpu_fphi with the trait command. The --plink option specifies a
plink data set that will determine which ids will be used in running the
EVD. The --use_covs options will include the ID set of covariates specified
through the covariate command.  This command now allows you to enter two
traits at once in order to get their corresponding ID set.

EXAMPLES:
Output consists of three files labeled:
\if{html}{\out{<output base filename>}}.ids --list of subject ids
\if{html}{\out{<output base filename>}}.eigenvalues --list of eigenvalues
\if{html}{\out{<output base filename>}}.eigenvectors --list of eigenvectors
\if{html}{\out{<output base filename>}}.notes --notes on the creation of the EVD data set
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateEvdData-new}{\code{CreateEvdData$new()}}
\item \href{#method-CreateEvdData-print}{\code{CreateEvdData$print()}}
\item \href{#method-CreateEvdData-finalize}{\code{CreateEvdData$finalize()}}
\item \href{#method-CreateEvdData-get_args}{\code{CreateEvdData$get_args()}}
\item \href{#method-CreateEvdData-clone}{\code{CreateEvdData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateEvdData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateEvdData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateEvdData$new(output_fpath = NULL, plink_fpath = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateEvdData-print"></a>}}
\if{latex}{\out{\hypertarget{method-CreateEvdData-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateEvdData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateEvdData-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CreateEvdData-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateEvdData$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateEvdData-get_args"></a>}}
\if{latex}{\out{\hypertarget{method-CreateEvdData-get_args}{}}}
\subsection{Method \code{get_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateEvdData$get_args()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateEvdData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateEvdData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateEvdData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
