% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarPar.R
\name{explainedVarProp}
\alias{explainedVarProp}
\title{Get explained variances for a group of SOLAR models.}
\usage{
explainedVarProp(mod)
}
\arguments{
\item{mod}{An object of \code{solarPolygenic} class. See \code{\link{solarPolygenicClass}}.}
}
\value{
A data frame with two columns \code{covariate} and \code{explainedVarProp}.
}
\description{
Get explained variances for a group of SOLAR models.
}
