% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{FPHI}
\alias{FPHI}
\title{R6 Class FPHI}
\description{
TODO: Add description
}
\details{
From the Solar manual:
Purpose: Fast test and heritability approximation

USAGE:
fphi 

\if{html}{\out{<div class="sourceCode">}}\preformatted{ -fast Performs a quick estimation run
 -debug Displays values at each iteration
 -list performs fast fphi on a list of trait (does not include covariate
   data)
 -precision number of decimals to calculate h2r
 -mask outputs fphi -fast results of the list of voxels from -list option
 -evd_data When using the -list option the EVD data option can be used
   to avoid having to calculate EVD data within the command
}\if{html}{\out{</div>}}

Fast permutation and heritability inference (FPHI). FPHI is based on the
eigenvalue decomposition on the kinship matrix and a search through values
of h2r for accurate approximation of heritability values and statistical
inference. The default setting is the full search out to 9 decimal places of
h2r. The "-fast" option uses the Wald approximation. Both use log likelyhood
p-value estimation.  The default setting should provide very accurate h2
estimates that are nearly identical to the standard maximum likelihood
inference. The h2 values obtained using Wald approximation are usually
within 3\% of the classical MLE values.  The same functionality is available
for GPU computing. Use gpu_fphi for heritability calculations in very large
datasets. For details see Ganjgahi et al., "Fast and powerful heritability
inference for family-based neuroimaging studies”.

EXAMPLES:
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FPHI-new}{\code{FPHI$new()}}
\item \href{#method-FPHI-print}{\code{FPHI$print()}}
\item \href{#method-FPHI-finalize}{\code{FPHI$finalize()}}
\item \href{#method-FPHI-get_opts}{\code{FPHI$get_opts()}}
\item \href{#method-FPHI-get_opts_fname}{\code{FPHI$get_opts_fname()}}
\item \href{#method-FPHI-get_precision}{\code{FPHI$get_precision()}}
\item \href{#method-FPHI-get_mask}{\code{FPHI$get_mask()}}
\item \href{#method-FPHI-get_evd_data}{\code{FPHI$get_evd_data()}}
\item \href{#method-FPHI-clone}{\code{FPHI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-new"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$new(
  opts = NULL,
  opts_fname = NULL,
  precision = NULL,
  mask = NULL,
  evd_data = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-print"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-get_opts"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-get_opts}{}}}
\subsection{Method \code{get_opts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$get_opts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-get_opts_fname"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-get_opts_fname}{}}}
\subsection{Method \code{get_opts_fname()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$get_opts_fname()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-get_precision"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-get_precision}{}}}
\subsection{Method \code{get_precision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$get_precision()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-get_mask"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-get_mask}{}}}
\subsection{Method \code{get_mask()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$get_mask()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-get_evd_data"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-get_evd_data}{}}}
\subsection{Method \code{get_evd_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$get_evd_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FPHI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FPHI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FPHI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
