% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSolarPolygenic.R
\name{getFormulaStr}
\alias{getFormulaStr}
\title{Get formula string from a solarPolygenic object}
\usage{
getFormulaStr(x)
}
\arguments{
\item{x}{An object of \code{solarPolygenic} object.}
}
\value{
A character string with formula.
}
\description{
The function returns a character string with formula.
The formula is derived based on \code{traits} and \code{covlist} slots of input object.
}
