% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{Load}
\alias{Load}
\title{R6 Class Load}
\description{
TODO: add description.
}
\details{
From the Solar manual:

USAGE: load \if{html}{\out{<object-type>}}  \if{html}{\out{<args>}}

EXAMPLES:
load pedigree    \if{html}{\out{<filename>}}
load phenotypes  \if{html}{\out{<filename>}}
load matrix      \link{-sample | -allow} \if{html}{\out{<filename>}} \if{html}{\out{<name1>}} 
load matrix       \if{html}{\out{<filename>}} ;# MathMatrix
load model       \if{html}{\out{<filename>}}
load freq        \link{-nosave} \if{html}{\out{<filename>}}
load marker      \link{-xlinked} \if{html}{\out{<filename>}}
load map         \link{-haldane | -kosambi} \if{html}{\out{<filename>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Load-new}{\code{Load$new()}}
\item \href{#method-Load-get_obj}{\code{Load$get_obj()}}
\item \href{#method-Load-get_opts}{\code{Load$get_opts()}}
\item \href{#method-Load-get_fpath}{\code{Load$get_fpath()}}
\item \href{#method-Load-get_cond}{\code{Load$get_cond()}}
\item \href{#method-Load-print}{\code{Load$print()}}
\item \href{#method-Load-finalize}{\code{Load$finalize()}}
\item \href{#method-Load-clone}{\code{Load$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-new"></a>}}
\if{latex}{\out{\hypertarget{method-Load-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$new(obj = NULL, opts = NULL, fpath = NULL, cond = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-get_obj"></a>}}
\if{latex}{\out{\hypertarget{method-Load-get_obj}{}}}
\subsection{Method \code{get_obj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$get_obj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-get_opts"></a>}}
\if{latex}{\out{\hypertarget{method-Load-get_opts}{}}}
\subsection{Method \code{get_opts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$get_opts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-get_fpath"></a>}}
\if{latex}{\out{\hypertarget{method-Load-get_fpath}{}}}
\subsection{Method \code{get_fpath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$get_fpath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-get_cond"></a>}}
\if{latex}{\out{\hypertarget{method-Load-get_cond}{}}}
\subsection{Method \code{get_cond()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$get_cond()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-print"></a>}}
\if{latex}{\out{\hypertarget{method-Load-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Load-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Load-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Load-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Load$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
