% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{load_strs_fmt}
\alias{load_strs_fmt}
\title{Format Load Commands}
\usage{
load_strs_fmt(private)
}
\arguments{
\item{private}{A reference to the private environment of the Solar class.}
}
\value{
A character vector of \code{load} command strings for use in TCL output.
}
\description{
Formats the \code{load} statements from internal \code{private$.loads} object.
}
