% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.lib.R
\name{loadExamplesPhen}
\alias{loadExamplesPhen}
\title{Load the example data (GAW10) from SOLAR.}
\usage{
loadExamplesPhen(dat.dir)
}
\arguments{
\item{dat.dir}{(optional) A character with path to a directory, where SOLAR files to be exported.
If this argument is missing, a temporary directory is used.}
}
\value{
A data frame returned by \code{readPhen} function called
with phen and ped files of the data set.
}
\description{
Function loads the example data distributed with SOLAR.
See the SOLAR tutorial for description of the data set.
}
