% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.lib.R
\name{loadMulticPhen}
\alias{loadMulticPhen}
\title{Load the complete data set from R package multic}
\usage{
loadMulticPhen()
}
\description{
The function loads the complete data of 12,000 individuals,
which is stored in .phen and .ped files.
These two files were generated within R package \code{multic}
and re-distributed in R package \code{solareclipser}
(\code{extdata/solarOutput} directory).
}
\details{
Function \code{\link{readPhen}} is used to read .phen and .ped files.
}
\examples{
dat <- loadMulticPhen()
dim(dat)

data(dat30)
dim(dat30)
}
