% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.lib.R
\name{matchIdNames}
\alias{matchIdNames}
\title{Match ID column names}
\usage{
matchIdNames(fields, sex.optional = FALSE, skip.sex = FALSE)
}
\arguments{
\item{fields}{The name of fields or colulm names in data frame, which are candidates to be ID fields.}

\item{sex.optional}{Logical, indicating if through an error in the case SEX field is not found in input fields.}

\item{skip.sex}{Logical, indicating if the search for SEX filds is completely skipped.}
}
\value{
A named character vector, that can be directly passed to \code{rename} function of \code{plyr} package.
}
\description{
The function automates the process of renaming ID fields in data frame of phenotypes
into a SOLAR format.
}
