% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.lib.R
\name{matchMapNames}
\alias{matchMapNames}
\title{Match map column names}
\usage{
matchMapNames(fields)
}
\arguments{
\item{fields}{The name of fields or colulm names in data frame, which are candidates to be map fields.}
}
\value{
A named character vector, that can be directly passed to \code{rename} function of \code{plyr} package.
}
\description{
The function searches for fields that correspond to those
in SOLAR format of marker map file.
}
\examples{
 data(dat50)
 matchMapNames(names(snpdata))
}
