% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.lib.R
\name{package.file}
\alias{package.file}
\title{Alternative to system.file}
\usage{
package.file(...)
}
\arguments{
\item{...}{arguments to be passed to \code{system.file}}
}
\value{
Path returned by \code{system.file}.
}
\description{
The function works as \code{system.file},
but takes care when the package is a local folder.
}
\details{
The use case is when some data file or direcotry is needed to be loaded,
and it is placed in \code{inst/} directory of the package.
}
\examples{
mibddir <- package.file("extdata", "solarOutput", "solarMibds", package = "solareclipser")
mibddir

list.files(mibddir)

}
