% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lib.R
\name{plotPed}
\alias{plotPed}
\title{Plot the pedigree}
\usage{
plotPed(data, ped, affected)
}
\arguments{
\item{data}{A data.frame of phenotypes.}

\item{ped}{An integer or a character value, that indicates the pedigree.}
}
\description{
Plot the pedigree tree based on the ID-fields given in phenotypes.
Plotting is based on \code{kinship2} package.
}
\details{
The ID fields are extracted from the data via \code{matchIdNames} function.
The required fields are ID, FA, MO, SEX and FAMID.
}
\examples{
data(dat30)
plotPed(dat30, 1)

}
\seealso{
\code{\link{matchIdNames}}
}
