% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{print_tcl_f_info}
\alias{print_tcl_f_info}
\title{Print TCL File Information}
\usage{
print_tcl_f_info(tcl_f_realpath, tcl_f_basename, tcl_proc_name)
}
\arguments{
\item{tcl_f_realpath}{Full path to TCL script.}

\item{tcl_f_basename}{Basename of TCL script.}

\item{tcl_proc_name}{Name of the TCL procedure.}
}
\value{
None. Prints output to console.
}
\description{
Utility to print full TCL file paths and metadata.
}
