% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{Solar}
\alias{Solar}
\title{R6 Class Solar}
\description{
This class is used to interact with \code{solar} and the files that are
created during analysis. The interface's methods and their arguments mimic
the \code{solar} CLI and their arguments when appropriate.
}
\details{
TODO: add more details
}
\examples{
\dontrun{

#==================== Ex. fphi() ====================
solar <- Solar$new()
solar$load(obj = "pedigree",
           fpath = "tests/input/solar/HCP_imputed_filtered_ped.csv",
           cond = "-t 0")
solar$load(obj = "phenotypes",
           fpath = "tests/input/solar/HCP_WM_ave_norm.csv")
solar$trait("CC")
solar$create_evd_data(output_fbasename = "evd_data")
solar$fphi(evd_data = "evd_data")
solar$run()

#==================== Ex. polygenic() ====================
solar <- Solar$new()
solar$load(obj = "pedigree",
           fpath = "tests/input/solar/HCP_imputed_filtered_ped.csv",
           cond = "-t 0")
solar$load(obj = "phenotypes",
           fpath = "tests/input/solar/HCP_WM_ave_norm.csv")
solar$trait("CC")$polygenic()
solar$run()

}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Solar-new}{\code{Solar$new()}}
\item \href{#method-Solar-print}{\code{Solar$print()}}
\item \href{#method-Solar-finalize}{\code{Solar$finalize()}}
\item \href{#method-Solar-set_run_rc}{\code{Solar$set_run_rc()}}
\item \href{#method-Solar-get_run_rc}{\code{Solar$get_run_rc()}}
\item \href{#method-Solar-set_settings}{\code{Solar$set_settings()}}
\item \href{#method-Solar-run}{\code{Solar$run()}}
\item \href{#method-Solar-clone}{\code{Solar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-new"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$new(settings = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-print"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-set_run_rc"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-set_run_rc}{}}}
\subsection{Method \code{set_run_rc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$set_run_rc(run_rc = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-get_run_rc"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-get_run_rc}{}}}
\subsection{Method \code{get_run_rc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$get_run_rc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-set_settings"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-set_settings}{}}}
\subsection{Method \code{set_settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$set_settings(settings = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-run"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
