% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solar_load_phenotype}
\alias{solar_load_phenotype}
\title{Load phenotype file}
\usage{
solar_load_phenotype(phenotype_filename)
}
\arguments{
\item{phenotype_filename}{Path to the phenotype CSV file}
}
\value{
\code{TRUE}, invisibly. Signals an error on failure.
}
\description{
Load a phenotype file for analysis. Pedigree must be loaded first.
}
