% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSolarAssoc.R
\name{solarAssocClass}
\alias{solarAssocClass}
\alias{print.solarAssoc}
\title{S3 class solarAssoc.}
\usage{
\method{print}{solarAssoc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{solarAssoc}.}

\item{...}{Additional arguments.}

\item{object}{An object of class \code{solarAssoc}.}

\item{y}{Character argument for \code{plot} method,
indicating the type of plot.
The default value is \code{"manh"}.}

\item{alpha}{Numeric argument between 0 and 1 for \code{summary} method,
indicating the significance level after Bonferroni multiple-test correction.
The default value is 0.05.}
}
\description{
S3 class solarAssoc.
}
