% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solareclipser-package.R
\docType{package}
\name{solareclipser-package}
\alias{solareclipser}
\alias{solareclipser-package}
\title{solareclipser: R Interface to 'SOLAR-Eclipse' Heritability Analysis}
\description{
An R interface to the Fast Permutation Heritability Inference (FPHI) analysis of 'SOLAR-Eclipse', ported from its C++ and Fortran source. Loads an empirical kinship pedigree and a phenotype file, adjusts a trait for covariates, and estimates its heritability (h2r) with standard errors and a likelihood ratio test, returning the results as data frames. Results can also be written in the format of the original 'SOLAR' fphi command. See \url{https://solar-eclipse-genetics.org} for more information about 'SOLAR'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/txbrain-org/solareclipser}
}

}
\author{
\strong{Maintainer}: William Burroughs \email{wilbura009@gmail.com}

Authors:
\itemize{
  \item William Burroughs \email{wilbura009@gmail.com}
}

}
\keyword{internal}
