% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSolarMultipoint.R
\name{solarMultipointClass}
\alias{solarMultipointClass}
\alias{print.solarMultipoint}
\title{S3 class solarMultipoint.}
\usage{
\method{print}{solarMultipoint}(x, ...)
}
\arguments{
\item{x}{An object of class \code{solarMultipoint}.}

\item{...}{Additional arguments.}

\item{object}{An object of class \code{solarMultipoint}.}

\item{pass}{Integer argument for \code{plot} method,
indicating whether which pass in multi-passs linkage scan to be plotted.
The default value is 1.}

\item{LOD.thr}{An argument for \code{tabplot} method.
The default value is \code{1.5}.}

\item{plot.null}{An argument for \code{tabplot} method.
The default value is \code{TRUE}.}
}
\description{
S3 class solarMultipoint.
}
