% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMultivar.R
\name{solarMultivarClass}
\alias{solarMultivarClass}
\alias{print.solarMultivar}
\title{S3 class solarMultivar.}
\usage{
\method{print}{solarMultivar}(x, ...)
}
\arguments{
\item{x}{An object of class \code{solarMultivar}.}

\item{...}{Additional arguments.}
}
\description{
S3 class solarMultivar.
}
