% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarPolyAssoc.R
\name{solarPolyAssoc}
\alias{solarPolyAssoc}
\alias{print.solarPolyAssoc}
\title{S3 class solarPolyAssoc.}
\usage{
\method{print}{solarPolyAssoc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{solarPolyAssoc}.}

\item{...}{Additional arguments.}
}
\description{
S3 class solarPolyAssoc.
}
