% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.lib.R
\name{solarReadFiles}
\alias{solarReadFiles}
\title{Read SOLAR output files in a directory}
\usage{
solarReadFiles(dir)
}
\arguments{
\item{dir}{A character, path to the directory.}
}
\value{
A list with file contents acquired by \code{readLines} function.
}
\description{
The function is used, for example, to read the files
in the output directory of the polygenic model.
}
