% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{tb_settings_set_output_dir}
\alias{tb_settings_set_output_dir}
\title{Return a list of settings for the testbench}
\usage{
tb_settings_set_output_dir(dir)
}
\arguments{
\item{dir}{The output directory to be used in the returned settings list.}
}
\value{
A named list containing testbench settings with structure \code{list(output = list(dir = ..., ...))}.
}
\description{
Constructs a \code{settings} list used for testing, setting the output directory and toggling other options (TCL logging, output redirection).
}
