% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{testbench_init}
\alias{testbench_init}
\title{Set common variables used test_that() tests
Returns: list(settings, pedigree_fpath, phenotypes_fpath)}
\usage{
testbench_init(tmp = FALSE)
}
\description{
Set common variables used test_that() tests
Returns: list(settings, pedigree_fpath, phenotypes_fpath)
}
