% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{Covariate}
\alias{Covariate}
\title{R6 Class Covariate}
\description{
TODO: add description.
}
\details{
From the Solar manual:

USAGE:
covariate \if{html}{\out{<variable>}}[^n | ^1,2\link{,3...}][\emph{\if{html}{\out{<variable>}} | #\if{html}{\out{<variable>}}
[(\link{trait})]]}
Creates a new covariate.  See below for examples.
;
covariate                      ; display all covariate info
covariate delete \if{html}{\out{<string>}}      ; deletes covariate and beta(s)
covariate delete_all           ; deletes all covariates and beta(s)
;
covariate \if{html}{\out{<variable>}}()         ; Null Covariate: require var in
;   sample without covariation
;
; Covariate Suspension (for
;   temporary hypothesis testing).
covariate suspend \if{html}{\out{<string>}}     ; temporarily disable covariate
covariate restore \if{html}{\out{<string>}}     ; re-activate suspended covariate

EXAMPLES:
covariate age                       ; Simple covariate Age
covariate age\emph{sex                   ; Age by Sex interaction (only)
covariate age}diabet\emph{diameds        ; 3-way interaction
covariate age^2                     ; Age squared as a simple covariate
covariate age^1,2                   ; Shorthand for: age age^2
covariate age#diabet                ; Shorthand for the following:
;   covariate age diabet age}diabet
covariate age^1,2,3#sex             ; Shorthand for all the following:
covariate sex age age\emph{sex age^2 age^2}sex age^3 age^3*sex

covariate sex age(q1) age\emph{sex(q3)   ; Trait-specific Covariates:
;   covariate sex applied to all traits
;   covariate age applied to trait q1
;   covariate age}sex applied to q3
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Covariate-new}{\code{Covariate$new()}}
\item \href{#method-Covariate-print}{\code{Covariate$print()}}
\item \href{#method-Covariate-finalize}{\code{Covariate$finalize()}}
\item \href{#method-Covariate-get_args}{\code{Covariate$get_args()}}
\item \href{#method-Covariate-clone}{\code{Covariate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Covariate-new"></a>}}
\if{latex}{\out{\hypertarget{method-Covariate-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$new(args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Covariate-print"></a>}}
\if{latex}{\out{\hypertarget{method-Covariate-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Covariate-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Covariate-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Covariate-get_args"></a>}}
\if{latex}{\out{\hypertarget{method-Covariate-get_args}{}}}
\subsection{Method \code{get_args()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$get_args()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Covariate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Covariate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
