% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{PedifromSNPs}
\alias{PedifromSNPs}
\title{R6 Class Pedifromsnps}
\description{
TODO: add description.
}
\details{
From the Solar manual:

Purpose: Creates a empirical pedigree matrix from a plink data set

USAGE:
pedifromsnps -i \if{html}{\out{<input base name of plink data>}}
-o \if{html}{\out{<output csv file name>}}
--freq \if{html}{\out{<file made with plink_freq>}}
OPTIONAL:
-corr \if{html}{\out{<alpha value>}}
-per-chromo -king -method_two -normalize
-batch_size \if{html}{\out{<batch size value>}}
-id_list <file w/ subject IDs>
-n_threads \if{html}{\out{<number of CPU threads>}}

DESCPRIPTION:
-i \if{html}{\out{<file>}}
The base file name of the plink .bed, .bim, and .fam files.
-o \if{html}{\out{<file>}}
The base file name for the output.
-freq \if{html}{\out{<file>}}
Name of output file from plink_freq command.
-n_threads \if{html}{\out{<n>}}
Number of CPU threads used for matrix calculation.
Default: Automatically set based on hardware
-per-chromo
Outputs a separate matrix for each chromosome.
Default: Disabled
-corr \if{html}{\out{<alpha value>}}
Compute method one correlation GRM using this alpha value.
Default: -1
-method_two
Computes correlation GRM using a second method described below.
Default: Disabled
-king
Computes Robust King GRM instead of using a correlation method.
Default: Disabled
-batch_size \if{html}{\out{<batch size value>}}
Number of loci computed at a single time per CPU thread.
Default: 500
-id_list <file w/ subject IDs>
Specified file contains a list of subject IDs separated by spaces.
The resulting GRM will only use these IDs and excluded all others.
Default: All IDs are used
-normalize
When used during the creation of a correlation GRM the final
values are normalized using the square roots of the diagonal values.
The result being that diagonal elements are 1 and off-diagonal
elements are bounded by 1 and -1.  Z*_i_j = Z_i_j/sqrt(Z_i_i\emph{Z_j_j)
where Z} is the final value and Z is the unnormalized value, i refers
to the index of subject i while j refers to the index of subject j.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PedifromSNPs-new}{\code{PedifromSNPs$new()}}
\item \href{#method-PedifromSNPs-print}{\code{PedifromSNPs$print()}}
\item \href{#method-PedifromSNPs-finalize}{\code{PedifromSNPs$finalize()}}
\item \href{#method-PedifromSNPs-get_input_fbase}{\code{PedifromSNPs$get_input_fbase()}}
\item \href{#method-PedifromSNPs-get_output_fbase}{\code{PedifromSNPs$get_output_fbase()}}
\item \href{#method-PedifromSNPs-get_freq_fbase}{\code{PedifromSNPs$get_freq_fbase()}}
\item \href{#method-PedifromSNPs-get_corr}{\code{PedifromSNPs$get_corr()}}
\item \href{#method-PedifromSNPs-get_per_chromo}{\code{PedifromSNPs$get_per_chromo()}}
\item \href{#method-PedifromSNPs-get_king}{\code{PedifromSNPs$get_king()}}
\item \href{#method-PedifromSNPs-get_method_two}{\code{PedifromSNPs$get_method_two()}}
\item \href{#method-PedifromSNPs-get_batch_size}{\code{PedifromSNPs$get_batch_size()}}
\item \href{#method-PedifromSNPs-get_id_list}{\code{PedifromSNPs$get_id_list()}}
\item \href{#method-PedifromSNPs-get_n_threads}{\code{PedifromSNPs$get_n_threads()}}
\item \href{#method-PedifromSNPs-clone}{\code{PedifromSNPs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-new"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$new(
  input_fbase = NULL,
  output_fbase = NULL,
  freq_fbase = NULL,
  corr = NULL,
  per_chromo = FALSE,
  king = FALSE,
  method_two = FALSE,
  batch_size = NULL,
  id_list = NULL,
  n_threads = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-print"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_input_fbase"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_input_fbase}{}}}
\subsection{Method \code{get_input_fbase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_input_fbase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_output_fbase"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_output_fbase}{}}}
\subsection{Method \code{get_output_fbase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_output_fbase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_freq_fbase"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_freq_fbase}{}}}
\subsection{Method \code{get_freq_fbase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_freq_fbase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_corr"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_corr}{}}}
\subsection{Method \code{get_corr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_corr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_per_chromo"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_per_chromo}{}}}
\subsection{Method \code{get_per_chromo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_per_chromo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_king"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_king}{}}}
\subsection{Method \code{get_king()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_king()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_method_two"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_method_two}{}}}
\subsection{Method \code{get_method_two()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_method_two()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_batch_size"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_batch_size}{}}}
\subsection{Method \code{get_batch_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_batch_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_id_list"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_id_list}{}}}
\subsection{Method \code{get_id_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_id_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-get_n_threads"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-get_n_threads}{}}}
\subsection{Method \code{get_n_threads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$get_n_threads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedifromSNPs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PedifromSNPs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedifromSNPs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
