% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{Polygenic}
\alias{Polygenic}
\title{R6 Class Polygenic}
\description{
TODO: Add description
}
\details{
From the Solar manual:

USAGE:
polygenic \link{-screen} \link{-all}  
 \link{-testrhoe} \link{-testrhog} \link{-testrhoc}
\link{-sporadic} \link{-keephouse} \link{-testrhop} \link{-rhopse} \link{-fphi}

(screencov is an alias for 'polygenic -screen')
(sporadic is an alias for 'polygenic -sporadic')

Typically before giving this command, you will give trait,
covariate, and house (if applicable) commands.  You will also load
pedigree and phenotypes files if they have not already been loaded.

EXAMPLES:
solar> load pedigree ped
solar> load phenotypes phen
solar> trait hbp
solar> covariate age sex age*sex smoke
solar> polygenic -screen
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Polygenic-new}{\code{Polygenic$new()}}
\item \href{#method-Polygenic-print}{\code{Polygenic$print()}}
\item \href{#method-Polygenic-finalize}{\code{Polygenic$finalize()}}
\item \href{#method-Polygenic-get_opts}{\code{Polygenic$get_opts()}}
\item \href{#method-Polygenic-clone}{\code{Polygenic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygenic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Polygenic-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygenic$new(opts = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygenic-print"></a>}}
\if{latex}{\out{\hypertarget{method-Polygenic-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygenic$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygenic-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Polygenic-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygenic$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygenic-get_opts"></a>}}
\if{latex}{\out{\hypertarget{method-Polygenic-get_opts}{}}}
\subsection{Method \code{get_opts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygenic$get_opts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygenic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Polygenic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygenic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
