% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{Solar}
\alias{Solar}
\title{R6 Class Solar}
\description{
TODO: add description.
}
\details{
Example solar command steps:
load phen HCP_WM_ave_norm.csv
trait CC
polygen
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Solar-new}{\code{Solar$new()}}
\item \href{#method-Solar-print}{\code{Solar$print()}}
\item \href{#method-Solar-finalize}{\code{Solar$finalize()}}
\item \href{#method-Solar-get_solar_command}{\code{Solar$get_solar_command()}}
\item \href{#method-Solar-set_solar_command}{\code{Solar$set_solar_command()}}
\item \href{#method-Solar-set_loads}{\code{Solar$set_loads()}}
\item \href{#method-Solar-get_loads}{\code{Solar$get_loads()}}
\item \href{#method-Solar-load}{\code{Solar$load()}}
\item \href{#method-Solar-trait}{\code{Solar$trait()}}
\item \href{#method-Solar-covariate}{\code{Solar$covariate()}}
\item \href{#method-Solar-polygenic}{\code{Solar$polygenic()}}
\item \href{#method-Solar-create_evd_data}{\code{Solar$create_evd_data()}}
\item \href{#method-Solar-fphi}{\code{Solar$fphi()}}
\item \href{#method-Solar-run}{\code{Solar$run()}}
\item \href{#method-Solar-clone}{\code{Solar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-new"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$new(save_output_dir = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-print"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$print(opts = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-get_solar_command"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-get_solar_command}{}}}
\subsection{Method \code{get_solar_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$get_solar_command()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-set_solar_command"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-set_solar_command}{}}}
\subsection{Method \code{set_solar_command()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$set_solar_command(solar_command)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-set_loads"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-set_loads}{}}}
\subsection{Method \code{set_loads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$set_loads(loads = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-get_loads"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-get_loads}{}}}
\subsection{Method \code{get_loads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$get_loads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-load"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$load(obj = NULL, opts = NULL, fpath = NULL, cond = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-trait"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-trait}{}}}
\subsection{Method \code{trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$trait(args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-covariate"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-covariate}{}}}
\subsection{Method \code{covariate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$covariate(covariate = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-polygenic"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-polygenic}{}}}
\subsection{Method \code{polygenic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$polygenic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-create_evd_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-create_evd_data}{}}}
\subsection{Method \code{create_evd_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$create_evd_data(
  output_fbasename = NULL,
  plink_fbasename = NULL,
  use_covs = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-fphi"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-fphi}{}}}
\subsection{Method \code{fphi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$fphi(
  opts = NULL,
  opts_fname = NULL,
  precision = NULL,
  mask = NULL,
  evd_data = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-run"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Solar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
