% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solar_select_trait}
\alias{solar_select_trait}
\title{Select trait for analysis}
\usage{
solar_select_trait(trait_name)
}
\arguments{
\item{trait_name}{Name of the trait column in the phenotype file}
}
\value{
Returns 0 on success, 1 on failure
}
\description{
Select a trait from the loaded phenotype file. Phenotypes must be loaded first.
}
