% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{SolarCommand}
\alias{SolarCommand}
\title{R6 Class Solar}
\description{
TODO: add description.
}
\details{
Example solar command steps:
load phen HCP_WM_ave_norm.csv
trait CC
polygen
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SolarCommand-new}{\code{SolarCommand$new()}}
\item \href{#method-SolarCommand-run}{\code{SolarCommand$run()}}
\item \href{#method-SolarCommand-about}{\code{SolarCommand$about()}}
\item \href{#method-SolarCommand-load}{\code{SolarCommand$load()}}
\item \href{#method-SolarCommand-trait}{\code{SolarCommand$trait()}}
\item \href{#method-SolarCommand-covariate}{\code{SolarCommand$covariate()}}
\item \href{#method-SolarCommand-polygenic}{\code{SolarCommand$polygenic()}}
\item \href{#method-SolarCommand-create_evd_data}{\code{SolarCommand$create_evd_data()}}
\item \href{#method-SolarCommand-get_load}{\code{SolarCommand$get_load()}}
\item \href{#method-SolarCommand-get_loads}{\code{SolarCommand$get_loads()}}
\item \href{#method-SolarCommand-get_trait}{\code{SolarCommand$get_trait()}}
\item \href{#method-SolarCommand-get_covariate}{\code{SolarCommand$get_covariate()}}
\item \href{#method-SolarCommand-get_polygenic}{\code{SolarCommand$get_polygenic()}}
\item \href{#method-SolarCommand-get_create_evd_data}{\code{SolarCommand$get_create_evd_data()}}
\item \href{#method-SolarCommand-get_save_output_dir}{\code{SolarCommand$get_save_output_dir()}}
\item \href{#method-SolarCommand-print}{\code{SolarCommand$print()}}
\item \href{#method-SolarCommand-print_private}{\code{SolarCommand$print_private()}}
\item \href{#method-SolarCommand-finalize}{\code{SolarCommand$finalize()}}
\item \href{#method-SolarCommand-clone}{\code{SolarCommand$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-new"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$new(save_output_dir = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-run"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-about"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-about}{}}}
\subsection{Method \code{about()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$about()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-load"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$load(obj = NULL, opts = NULL, fpath = NULL, cond = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-trait"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-trait}{}}}
\subsection{Method \code{trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$trait(args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-covariate"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-covariate}{}}}
\subsection{Method \code{covariate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$covariate(covariate = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-polygenic"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-polygenic}{}}}
\subsection{Method \code{polygenic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$polygenic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-create_evd_data"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-create_evd_data}{}}}
\subsection{Method \code{create_evd_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$create_evd_data(output_fpath = NULL, plink_fpath = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_load"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_load}{}}}
\subsection{Method \code{get_load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_loads"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_loads}{}}}
\subsection{Method \code{get_loads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_loads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_trait"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_trait}{}}}
\subsection{Method \code{get_trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_trait()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_covariate"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_covariate}{}}}
\subsection{Method \code{get_covariate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_covariate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_polygenic"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_polygenic}{}}}
\subsection{Method \code{get_polygenic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_polygenic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_create_evd_data"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_create_evd_data}{}}}
\subsection{Method \code{get_create_evd_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_create_evd_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_save_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_save_output_dir}{}}}
\subsection{Method \code{get_save_output_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_save_output_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-print"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-print_private"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-print_private}{}}}
\subsection{Method \code{print_private()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$print_private(field = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
