% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarCommand.R
\name{SolarCommand}
\alias{SolarCommand}
\title{R6 Class SolarCommand}
\description{
TODO: add description.
}
\details{
Example solar command steps:
load phen HCP_WM_ave_norm.csv
trait CC
polygen
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SolarCommand-new}{\code{SolarCommand$new()}}
\item \href{#method-SolarCommand-load}{\code{SolarCommand$load()}}
\item \href{#method-SolarCommand-trait}{\code{SolarCommand$trait()}}
\item \href{#method-SolarCommand-polygenic}{\code{SolarCommand$polygenic()}}
\item \href{#method-SolarCommand-covariate}{\code{SolarCommand$covariate()}}
\item \href{#method-SolarCommand-create_evd_data}{\code{SolarCommand$create_evd_data()}}
\item \href{#method-SolarCommand-fphi}{\code{SolarCommand$fphi()}}
\item \href{#method-SolarCommand-pedifromsnps}{\code{SolarCommand$pedifromsnps()}}
\item \href{#method-SolarCommand-get_load}{\code{SolarCommand$get_load()}}
\item \href{#method-SolarCommand-get_loads}{\code{SolarCommand$get_loads()}}
\item \href{#method-SolarCommand-get_trait}{\code{SolarCommand$get_trait()}}
\item \href{#method-SolarCommand-get_covariate}{\code{SolarCommand$get_covariate()}}
\item \href{#method-SolarCommand-get_polygenic}{\code{SolarCommand$get_polygenic()}}
\item \href{#method-SolarCommand-get_create_evd_data}{\code{SolarCommand$get_create_evd_data()}}
\item \href{#method-SolarCommand-get_fphi}{\code{SolarCommand$get_fphi()}}
\item \href{#method-SolarCommand-get_pedifromsnps}{\code{SolarCommand$get_pedifromsnps()}}
\item \href{#method-SolarCommand-print}{\code{SolarCommand$print()}}
\item \href{#method-SolarCommand-finalize}{\code{SolarCommand$finalize()}}
\item \href{#method-SolarCommand-clone}{\code{SolarCommand$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-new"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-load"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$load(obj = NULL, opts = NULL, fpath = NULL, cond = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-trait"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-trait}{}}}
\subsection{Method \code{trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$trait(args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-polygenic"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-polygenic}{}}}
\subsection{Method \code{polygenic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$polygenic(opts = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-covariate"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-covariate}{}}}
\subsection{Method \code{covariate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$covariate(args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-create_evd_data"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-create_evd_data}{}}}
\subsection{Method \code{create_evd_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$create_evd_data(
  output_fbasename = NULL,
  plink_fbasename = NULL,
  use_covs = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-fphi"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-fphi}{}}}
\subsection{Method \code{fphi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$fphi(
  opts = NULL,
  opts_fname = NULL,
  precision = NULL,
  mask = NULL,
  evd_data = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-pedifromsnps"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-pedifromsnps}{}}}
\subsection{Method \code{pedifromsnps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$pedifromsnps(
  input_fbase = NULL,
  output_fbase = NULL,
  freq_fbase = NULL,
  corr = NULL,
  per_chromo = FALSE,
  king = FALSE,
  method_two = FALSE,
  batch_size = NULL,
  id_list = NULL,
  n_threads = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_load"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_load}{}}}
\subsection{Method \code{get_load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_loads"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_loads}{}}}
\subsection{Method \code{get_loads()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_loads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_trait"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_trait}{}}}
\subsection{Method \code{get_trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_trait()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_covariate"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_covariate}{}}}
\subsection{Method \code{get_covariate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_covariate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_polygenic"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_polygenic}{}}}
\subsection{Method \code{get_polygenic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_polygenic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_create_evd_data"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_create_evd_data}{}}}
\subsection{Method \code{get_create_evd_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_create_evd_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_fphi"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_fphi}{}}}
\subsection{Method \code{get_fphi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_fphi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-get_pedifromsnps"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-get_pedifromsnps}{}}}
\subsection{Method \code{get_pedifromsnps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$get_pedifromsnps()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-print"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SolarCommand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarCommand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
