% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solareclipser-package.R
\docType{package}
\name{solareclipser-package}
\alias{solareclipser}
\alias{solareclipser-package}
\title{solareclipser: solareclipser is an R package version of solareclipse}
\description{
'SOLAR' is the standard software program to perform linkage and association mappings of the quantitative trait loci (QTLs) in pedigrees of arbitrary size and complexity. This package allows the user to exploit the variance component methods implemented in 'SOLAR'. It automates such routine operations as formatting pedigree and phenotype data. It also parses the model output and contains summary and plotting functions for exploration of the results. In addition, 'solareclipser' enables parallel computing of the linkage and association analyses, that makes the calculation of genome-wide scans more efficient. See \url{https://solar-eclipse-genetics.org} for more information about 'SOLAR'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/txbrain-org/solareclipser}
}

}
\author{
\strong{Maintainer}: William Burroughs \email{wilbura009@gmail.com}

}
\keyword{internal}
