% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarFiles.R
\name{SolarFiles}
\alias{SolarFiles}
\title{R6 Class SolarFiles}
\description{
TODO: add description
}
\details{
TODO: add list of files created by solar
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SolarFiles-new}{\code{SolarFiles$new()}}
\item \href{#method-SolarFiles-get_solar_executable}{\code{SolarFiles$get_solar_executable()}}
\item \href{#method-SolarFiles-print}{\code{SolarFiles$print()}}
\item \href{#method-SolarFiles-clone}{\code{SolarFiles$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFiles-new"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFiles-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFiles$new(solar_executable = NA, solar_output_dir = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFiles-get_solar_executable"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFiles-get_solar_executable}{}}}
\subsection{Method \code{get_solar_executable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFiles$get_solar_executable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFiles-print"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFiles-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFiles$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFiles-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFiles-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFiles$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
