% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarFiles.R
\name{SolarFilesController}
\alias{SolarFilesController}
\title{R6 Class SolarFilesController}
\description{
TODO: add description
}
\details{
TODO: controller for SolarFiles
}
\section{Super class}{
\code{\link[solareclipser:SolarFiles]{solareclipser::SolarFiles}} -> \code{SolarFilesController}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SolarFilesController-new}{\code{SolarFilesController$new()}}
\item \href{#method-SolarFilesController-print}{\code{SolarFilesController$print()}}
\item \href{#method-SolarFilesController-set_polygenic_trait_d_files}{\code{SolarFilesController$set_polygenic_trait_d_files()}}
\item \href{#method-SolarFilesController-get_polygenic_trait_d_files}{\code{SolarFilesController$get_polygenic_trait_d_files()}}
\item \href{#method-SolarFilesController-clone}{\code{SolarFilesController$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="get_fphi_files"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-get_fphi_files'><code>solareclipser::SolarFiles$get_fphi_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="get_path_to_files"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-get_path_to_files'><code>solareclipser::SolarFiles$get_path_to_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="get_polygenic_files"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-get_polygenic_files'><code>solareclipser::SolarFiles$get_polygenic_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="get_shared_files"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-get_shared_files'><code>solareclipser::SolarFiles$get_shared_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_fphi_eigenvalues_f"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_fphi_eigenvalues_f'><code>solareclipser::SolarFiles$set_fphi_eigenvalues_f()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_fphi_eigenvectors_f"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_fphi_eigenvectors_f'><code>solareclipser::SolarFiles$set_fphi_eigenvectors_f()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_fphi_evd_data_bn"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_fphi_evd_data_bn'><code>solareclipser::SolarFiles$set_fphi_evd_data_bn()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_fphi_ids_f"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_fphi_ids_f'><code>solareclipser::SolarFiles$set_fphi_ids_f()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_fphi_notes_f"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_fphi_notes_f'><code>solareclipser::SolarFiles$set_fphi_notes_f()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_polygenic"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_polygenic'><code>solareclipser::SolarFiles$set_polygenic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solareclipser" data-topic="SolarFiles" data-id="set_polygenic_trait_d"><a href='../../solareclipser/html/SolarFiles.html#method-SolarFiles-set_polygenic_trait_d'><code>solareclipser::SolarFiles$set_polygenic_trait_d()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFilesController-new"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFilesController-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFilesController$new(sf = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFilesController-print"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFilesController-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFilesController$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFilesController-set_polygenic_trait_d_files"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFilesController-set_polygenic_trait_d_files}{}}}
\subsection{Method \code{set_polygenic_trait_d_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFilesController$set_polygenic_trait_d_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFilesController-get_polygenic_trait_d_files"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFilesController-get_polygenic_trait_d_files}{}}}
\subsection{Method \code{get_polygenic_trait_d_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFilesController$get_polygenic_trait_d_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarFilesController-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SolarFilesController-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarFilesController$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
