% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{tb_settings_set_output_dir}
\alias{tb_settings_set_output_dir}
\title{Return a list of settings for the testbench}
\usage{
tb_settings_set_output_dir(dir)
}
\arguments{
\item{dir}{The output directory to be used by the returned settings}
}
\value{
A list of settings
}
\description{
Return a list of settings for the testbench
}
